requirements/base.txt
Django==3.0.6
django-allauth==0.41.0
django-crispy-forms==1.9.0
django-environ==0.4.5
djangorestframework==3.11.0
django-emoji-picker==0.0.6